pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "veritensor"
version = "1.6.3"
description = "Antivirus for the AI Supply Chain. Scans models, datasets, notebooks, and RAG documents for threats."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
  {name = "Arsenii Brazhnyk", email = "veritensor@gmail.com"} 
]
keywords = ["security", "ai", "mlops", "malware-detection", "supply-chain", "devsecops", "rag", "data-poisoning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "huggingface_hub>=0.19.0",
    "cyclonedx-python-lib>=6.0.0",
    # Fickling removed from here to avoid installing PyTorch (800MB+) for regular users
]


[project.optional-dependencies]
# Install these only for testing: pip install veritensor[test]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "fickling>=0.0.1",
    "reportlab>=4.0.0",  
    "anyio>=4.0.0",      
    "coverage>=7.0.0"      
]

# Optional dependencies for PII scanning 
pii = [
    "presidio-analyzer>=2.2.0",
    "spacy>=3.0.0"
]

# Optional dependencies for RAG document scanning (PDF, DOCX)
rag = [
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.21",
    "veritensor[pii]"
]

# AWS Support (S3 scanning)
aws = [
    "boto3>=1.28.0"
]

data = [
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0"
]

all = [
    "veritensor[aws,data,rag,pii]"
]

dev = [
    "veritensor[all,test]", 
    "black",        
    "isort"
]

[project.urls]
"Homepage" = "https://github.com/arsbr/Veritensor"
"Bug Tracker" = "https://github.com/arsbr/Veritensor/issues"
"Documentation" = "https://github.com/arsbr/Veritensor#readme"

[project.scripts]
veritensor = "veritensor.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
veritensor = ["**/*.yaml"]
